requirements_freeze.txt
# Core
numpy==1.26.4
pandas==2.3.3
scikit-learn==1.5.2
scipy==1.16.3

# Config
hydra-core==1.3.2
omegaconf==2.3.0
PyYAML==6.0.3

# DL
torch==2.6.0
torchvision==0.21.0
torchaudio==2.6.0

# CV / IO
opencv-python-headless==4.10.0.84
pillow==12.0.0

# Logging
wandb==0.23.1

# Visualization 
matplotlib==3.8.4
seaborn==0.13.2

# Imbalance utils
imbalanced-learn==0.12.4

# Common utils 
requests==2.32.5
protobuf==5.29.5
psutil==7.1.3
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2
packaging==25.0

# Git-based experiment utilities 
GitPython==3.1.45

